requirements.txt
# ============================================================
# AI Bug Triage System — Pinned Dependencies
# Generated from pyproject.toml. Prefer `pip install -e ".[dev]"`
# ============================================================

# --- AI / LLM ---
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.0.0
tiktoken>=0.5.0

# --- UI ---
streamlit>=1.30.0

# --- Vector Databases ---
faiss-cpu>=1.7.4
chromadb>=0.4.0

# --- Data Models & Config ---
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# --- Data Processing ---
pandas>=2.0.0

# --- Dev Tools ---
pytest>=7.0.0
pytest-cov>=4.0.0
ruff>=0.1.0
mypy>=1.0.0
ipykernel>=6.0.0
jupyter>=1.0.0
